refactor: move markdown component
